split webpack configs for example
