+++ package.json
@@ -23,7 +23,7 @@
     "test:watch": "cross-env NODE_ENV=test jest --no-cache --watchAll",
     "analyze": "webpack --config internals/webpack/webpack.prod.babel.js --profile --json > stats.json",
     "example:build": "cd example && webpack-cli --progress",
-    "example:dev": "cd example && webpack-cli --progress --watch",
+    "example:dev": "cd example && webpack-cli --config webpack.config.dev.babel.js --progress --watch",
     "example:clean": "rm -rf example/dist"
   },
   "dllPlugin": {
@@ -109,13 +109,16 @@
     "localStorage": "1.0.3",
     "mini-css-extract-plugin": "^0.5.0",
     "null-loader": "0.1.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
     "preact-jsx-chai": "2.2.1",
     "prettier": "1.12.1",
     "shelljs": "0.8.2",
+    "uglifyjs-webpack-plugin": "^2.0.1",
     "url-loader": "1.0.1",
     "webpack": "4.8.3",
     "webpack-cli": "2.1.3",
     "webpack-dev-middleware": "3.1.3",
-    "webpack-hot-middleware": "2.22.2"
+    "webpack-hot-middleware": "2.22.2",
+    "webpack-merge": "^4.1.5"
   }
 }
